разбитие на модули
